requirements.txt
cffi==1.17.1
colorzero==2.0
gpiozero==2.0.1
lgpio==0.2.2.0
numpy==2.3.1
paho-mqtt==2.1.0
pycparser==2.22
pygame==2.6.1
pyserial==3.5
schedule==1.2.2
scipy==1.15.3
sounddevice==0.5.2
